Windows compatibility
Added the tzinfo-data gem so that this server page will run on Windows.
Apparently on OS X systems, the time zone info can be pulled directly
from the OS itself. Windows requires this additional gem to pull
timezone data. Without this gem you will encounter an error if
attempting to run this server on Windows.

Also ensured compatibility with Ruby 2.1.

P.S.
Matthew (or anyone using a Mac), please confirm that this server page
runs on OS X as well. No reason it shouldn't, as nothing was removed,
but it's always possible.